refactor: move the MP3 frame parsing to the MP3 parser instead of the tags
