+++ lib/src/parsers/tags/id3v2.dart
@@ -206,142 +206,7 @@ class ID3v2Parser extends TagParser {
       processFrame(frame.id, frame.size);
     }
 
-    if (metadata.duration == null || metadata.duration == Duration.zero) {
-      buffer.setPositionSync(size + 10);
-
-      final mp3FrameHeader = _findFirstMp3Frame(buffer);
-
-      if (mp3FrameHeader == null) {
-        return metadata;
-      }
-
-      final mpegVersion = switch ((mp3FrameHeader[1] >> 3) & 0x3) {
-        0x00 => 3,
-        0x01 => -1,
-        0x02 => 2,
-        0x03 => 1,
-        _ => -1
-      };
-      final mpegLayer = switch ((mp3FrameHeader[1] >> 1) & 0x3) {
-        0 => -1,
-        1 => 3,
-        2 => 2,
-        3 => 1,
-        _ => -1,
-      };
-
-      final bitrateIndex = mp3FrameHeader[2] >> 4;
-      final samplerateIndex = mp3FrameHeader[2] & 12 >> 0x3;
-
-      metadata.samplerate = _getSampleRate(mpegVersion, samplerateIndex);
-      metadata.bitrate = _getBitrate(mpegVersion, mpegLayer, bitrateIndex);
-
-      // arbitrary choice.  Usually the `Xing` header is located after ~30 bytes
-      // then the header size is about ~150 bytes.
-      final possibleXingHeader = buffer.readAtMost(1500);
-
-      int i = 0;
-      while (i < possibleXingHeader.length && possibleXingHeader[i] == 0) {
-        i++;
-      }
-
-      if ((i < possibleXingHeader.length - 11) &&
-          possibleXingHeader[i] == 0x58 &&
-          possibleXingHeader[i + 1] == 0X69 &&
-          possibleXingHeader[i + 2] == 0x6E &&
-          possibleXingHeader[i + 3] == 0x67) {
-        // it's a VBR file (Variable Bit Rate)
-        final xingFrameFlag = possibleXingHeader[i + 7] & 0x1;
-
-        if (xingFrameFlag == 1) {
-          final numberOfFrames =
-              getUint32(possibleXingHeader.sublist(i + 8, i + 12));
-          final samplesPerFrame =
-              _getSamplePerFrame(mpegVersion, mpegLayer) ?? 0;
-          final sampleRate = metadata.samplerate;
-
-          if (sampleRate != null && sampleRate > 0 && samplesPerFrame > 0) {
-            final totalSamples = numberOfFrames * samplesPerFrame;
-            final durationInSeconds = totalSamples / sampleRate;
-
-            final durationInMicroseconds =
-                (durationInSeconds * 1000000).toInt();
-            metadata.duration = Duration(microseconds: durationInMicroseconds);
-          }
-        }
-      } else {
-        // it's a CBR file (Constant Bit Rate)
-        if (metadata.bitrate != null && metadata.bitrate! > 0) {
-          final fileSizeWithoutMetadata = reader.lengthSync() - size;
-          final durationInSeconds =
-              (8 * fileSizeWithoutMetadata) / metadata.bitrate!;
-
-          // Convert to microseconds
-          final durationInMicroseconds = (durationInSeconds * 1000000).toInt();
-          metadata.duration = Duration(microseconds: durationInMicroseconds);
-        }
-      }
-    }
-
     return metadata;
-  }
-
-  /// Search and return the first MP3 frame header.
-  /// Returns null if none has been found.
-  ///
-  /// The MP3 frame has a magic word : 0xFFF or 0xFFE
-  ///
-  /// Sometimes the MP3 files contains blocks of 0x00 or 0xFF and relying on the magic word
-  /// is not reliable anymore.
-  ///
-  /// To prevent false positives, we need to verify that the bytes after the potential
-  /// valid word are correct. The MP3 specs specify several flags that must be set or not.
-  ///
-  /// Credit to [exiftool](https://github.com/exiftool/exiftool/blob/master/lib/Image/ExifTool/MPEG.pm#L464)
-  Uint8List? _findFirstMp3Frame(Buffer buffer) {
-    Uint8List frameHeader = buffer.readAtMost(4);
-
-    while (frameHeader.length == 4) {
-      // Look for frame sync (0xFF followed by 3 bytes)
-      if (frameHeader[0] == 0xFF) {
-        int word = (frameHeader[0] << 24) |
-            (frameHeader[1] << 16) |
-            (frameHeader[2] << 8) |
-            (frameHeader[3]);
-
-        if ((word & 0xFFE00000) != 0xFFE00000) {
-          frameHeader[0] = frameHeader[1];
-          frameHeader[1] = frameHeader[2];
-          frameHeader[2] = frameHeader[3];
-          frameHeader[3] = buffer.read(1)[0];
-          continue;
-        }
-
-        // Check for invalid MPEG version (01), layer (00), bitrate index (0000 or 1111),
-        // reserved sampling frequency (11), reserved emphasis (10), and not Layer III if MP3
-        if ((word & 0x180000) == 0x080000 || // reserved version ID
-            (word & 0x060000) == 0x000000 || // reserved layer
-            (word & 0x00F000) == 0x000000 || // free bitrate
-            (word & 0x00F000) == 0x00F000 || // bad bitrate
-            (word & 0x000C00) == 0x000C00 || // reserved sampling rate
-            (word & 0x000003) == 0x000002) {
-          frameHeader[0] = frameHeader[1];
-          frameHeader[1] = frameHeader[2];
-          frameHeader[2] = frameHeader[3];
-          frameHeader[3] = buffer.read(1)[0];
-          continue;
-        }
-
-        return frameHeader;
-      }
-
-      frameHeader[0] = frameHeader[1];
-      frameHeader[1] = frameHeader[2];
-      frameHeader[2] = frameHeader[3];
-      frameHeader[3] = buffer.read(1)[0];
-    }
-
-    return null;
   }
 
   /// Process a frame.
@@ -724,164 +589,5 @@ class ID3v2Parser extends TagParser {
     } else {
       return int.tryParse(year);
     }
-  }
-
-  int? _getSampleRate(int mpegVersion, int sampleRateIndex) {
-    if (mpegVersion == 1) {
-      return switch (sampleRateIndex) {
-        0 => 44100,
-        1 => 48000,
-        2 => 32000,
-        _ => null,
-      };
-    }
-
-    if (mpegVersion == 2) {
-      return switch (sampleRateIndex) {
-        0 => 22050,
-        1 => 24000,
-        2 => 16000,
-        _ => null,
-      };
-    }
-
-    if (mpegVersion == 3) {
-      return switch (sampleRateIndex) {
-        0 => 11025,
-        1 => 12000,
-        2 => 8000,
-        _ => null,
-      };
-    }
-
-    return null;
-  }
-
-  int? _getBitrate(int mpegVersion, int mpegLayer, int bitrateIndex) {
-    if (mpegVersion == 1 && mpegLayer == 1) {
-      return switch (bitrateIndex) {
-        0 => null,
-        1 => 32000,
-        2 => 64000,
-        3 => 96000,
-        4 => 128000,
-        5 => 160000,
-        6 => 192000,
-        7 => 224000,
-        8 => 256000,
-        9 => 288000,
-        10 => 320000,
-        11 => 352000,
-        12 => 384000,
-        13 => 416000,
-        14 => 448000,
-        _ => null,
-      };
-    }
-
-    if (mpegVersion == 1 && mpegLayer == 2) {
-      return switch (bitrateIndex) {
-        0 => null,
-        1 => 32000,
-        2 => 48000,
-        3 => 56000,
-        4 => 64000,
-        5 => 80000,
-        6 => 96000,
-        7 => 112000,
-        8 => 128000,
-        9 => 160000,
-        10 => 192000,
-        11 => 224000,
-        12 => 256000,
-        13 => 320000,
-        14 => 384000,
-        _ => null,
-      };
-    }
-
-    if (mpegVersion == 1 && mpegLayer == 3) {
-      return switch (bitrateIndex) {
-        0 => null,
-        1 => 32000,
-        2 => 40000,
-        3 => 48000,
-        4 => 56000,
-        5 => 64000,
-        6 => 80000,
-        7 => 96000,
-        8 => 112000,
-        9 => 128000,
-        10 => 160000,
-        11 => 192000,
-        12 => 224000,
-        13 => 256000,
-        14 => 320000,
-        _ => null,
-      };
-    }
-    if (mpegVersion == 2 && mpegLayer == 1) {
-      return switch (bitrateIndex) {
-        0 => null,
-        1 => 32000,
-        2 => 48000,
-        3 => 56000,
-        4 => 64000,
-        5 => 80000,
-        6 => 96000,
-        7 => 112000,
-        8 => 128000,
-        9 => 144000,
-        10 => 160000,
-        11 => 176000,
-        12 => 192000,
-        13 => 224000,
-        14 => 256000,
-        _ => null,
-      };
-    }
-
-    if (mpegVersion == 2 && (mpegLayer == 2 || mpegLayer == 3)) {
-      return switch (bitrateIndex) {
-        0 => null,
-        1 => 8000,
-        2 => 16000,
-        3 => 24000,
-        4 => 32000,
-        5 => 40000,
-        6 => 48000,
-        7 => 56000,
-        8 => 64000,
-        9 => 80000,
-        10 => 96000,
-        11 => 112000,
-        12 => 128000,
-        13 => 144000,
-        14 => 160000,
-        _ => null,
-      };
-    }
-
-    return null;
-  }
-
-  int? _getSamplePerFrame(int mpegAudioVersion, int mpegLayer) {
-    if (mpegAudioVersion == 1) {
-      return switch (mpegLayer) {
-        1 => 384,
-        2 => 1152,
-        3 => 1152,
-        _ => null,
-      };
-    } else if (mpegAudioVersion == 2) {
-      return switch (mpegLayer) {
-        1 => 192,
-        2 => 1152,
-        3 => 576,
-        _ => null,
-      };
-    }
-
-    return null;
   }
 }
